Split out PredictProvider in preparation for decoupling it from Predict
